replace short to int in Border operator

+++ simonov.konstantin/T2/Io.cpp
@@ -12,7 +12,7 @@ std::istream& simonov::operator>>(std::istream& in, Border&& fence)
   }
 
   std::string input;
-  short i = 0;
+  int i = 0;
   while (i != fence.length_)
   {
     char symbol;
